Cambios en los 4 Web service
1. HELP.md borrado.
2. Atributos int cambiados a long.

+++ accounting-management/src/main/java/com/nmatute/octoger/accountingmanagement/domain/dao/IProductCollectionRepository.java
@@ -8,5 +8,5 @@ import com.nmatute.octoger.accountingmanagement.domain.dto.ProductCollectionDTO;
  * @author NM4TT
  */
 public interface IProductCollectionRepository {
-    ProductCollectionDTO getById(int id);
+    ProductCollectionDTO getById(long id);
 }
